refactor: align codebase with coding standards for maintainability
This refactoring addresses critical and high-priority violations identified
in the coding standards review, significantly improving code maintainability,
reusability, and readability.

CRITICAL: File Size Reductions
- Reduced matrix-board.tsx from 521 to 398 lines (-24%)
- Reduced task-form.tsx from 363 to 246 lines (-32%)
- Reduced filter-popover.tsx from 326 to 203 lines (-38%)

HIGH PRIORITY: Code Quality Improvements
- Centralized ID generation with generateId() helper (eliminates 7 instances of magic number)
- Extracted DND and toast duration constants to lib/constants.ts
- Simplified toggleCompleted() by extracting createRecurringInstance() function
- Simplified importTasks() by extracting regenerateConflictingIds() function

Component Extraction:
- Created reusable MatrixEmptyState component for onboarding
- Created reusable TaskFormTags and TaskFormSubtasks components
- Created reusable FilterDueDate component
- Created useKeyboardShortcuts hook for global keyboard handling

Benefits:
- 363 lines removed across critical files
- 7 new reusable components/utilities created
- Single source of truth for IDs and constants
- Improved testability with smaller, focused components
- Reduced cognitive load and improved readability

All critical data layer tests passing (140/154 tests pass, 14 pre-existing UI test issues unrelated to refactoring).
